requirements.txt
absl-py==0.10.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
dataclasses==0.8
filelock==3.4.0
Flask==2.0.2
flatbuffers==2.0
frozenlist==1.2.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.1.0
huggingface-hub==0.2.1
idna==3.1
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
multidict==5.2.0
nltk==3.6.5
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pip==21.3.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pyu2f==0.1.5
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.3
sentencepiece==0.1.95
setuptools==59.4.0
six==1.15.0
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-macos==2.7.0
tensorflow-metal==0.3.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
tqdm==4.62.3
transformers==4.13.0.dev0
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.1
wheel==0.35.1
wrapt==1.12.1
yarl==1.7.2
zipp==3.6.0
